ASA-1694: support Authentication for username and password
